// dotnet/openapi/model/clientCompanyModelListPagedResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClientCompanyModel } from './clientCompanyModel';


export interface ClientCompanyModelListPagedResponse { 
    message?: string | null;
    statusCode?: number;
    data?: Array<ClientCompanyModel> | null;
    pageIndex?: number;
    pageSize?: number;
    totalRecords?: number;
}